livekit-agent-server/speakingreport_update/requirements.txt
aiohttp==3.11.18
Flask==3.1.0
livekit==1.0.6
livekit_api==1.0.2
livekit_plugins_noise_cancellation==0.2.1
livekit_protocol==1.0.2
openai==1.77.0
python-dotenv==1.1.0
